backend/requirements.txt
# GitAI Coding Agent — Python dependencies
# Install: pip install -r requirements.txt

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# HTTP client (async, streaming SSE support)
httpx>=0.27.0

# Environment + config
python-dotenv>=1.0.0

# Prompt templating
jinja2>=3.1.0

# Async file I/O
aiofiles>=23.2.0

# CLI
click>=8.1.0
rich>=13.7.0

# .gitignore parsing
pathspec>=0.12.0

# Pydantic (FastAPI models)
pydantic>=2.0.0

# GitHub OAuth (existing web UI feature)
requests>=2.31.0
